export Deterministic wallet logic into single component and apply it on wallets
