+++ src/components/WalletUnlock/Ledger.tsx
@@ -2,18 +2,10 @@ import React, { useContext, useState } from 'react';
 import uniqBy from 'ramda/src/uniqBy';
 import prop from 'ramda/src/prop';
 
-import { MOONPAY_ASSET_UUIDS, IS_ELECTRON, accountsToCSV } from '@utils';
+import { MOONPAY_ASSET_UUIDS, IS_ELECTRON } from '@utils';
 import { FormData, WalletId, ExtendedAsset } from '@types';
 import translate, { translateRaw, Trans } from '@translations';
-import {
-  NewTabLink,
-  Spinner,
-  Button,
-  DeterministicAccountList,
-  AssetSelector,
-  Typography,
-  Input
-} from '@components';
+import { NewTabLink, Spinner, Button } from '@components';
 import {
   EXT_URLS,
   LEDGER_DERIVATION_PATHS,
@@ -32,68 +24,12 @@ import {
 
 import ledgerIcon from '@assets/images/icn-ledger-nano-large.svg';
 import UnsupportedNetwork from './UnsupportedNetwork';
-import styled from 'styled-components';
-import { COLORS, SPACING } from '@theme';
-import { Formik } from 'formik';
-import Icon from '@components/Icon';
+import DeterministicWallet from './DeterministicWallet';
 
 interface OwnProps {
   formData: FormData;
   onUnlock(param: any): void;
 }
-
-const MnemonicWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
-
-const Title = styled(Typography)`
-  color: ${COLORS.BLUE_DARK};
-  margin-bottom: 20px;
-`;
-
-const Parameters = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  margin-bottom: 40px;
-  margin-top: 15px;
-`;
-
-const TableContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`;
-
-const HeadingWrapper = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-`;
-
-const SForm = styled.form`
-  width: 50%;
-`;
-
-const SLabel = styled.label`
-  font-weight: normal;
-  margin-bottom: ${SPACING.SM};
-  margin-top: ${SPACING.MD};
-`;
-
-const SButton = styled(Button)`
-  margin: 30px 0;
-`;
-
-const SLink = styled.span`
-  color: ${COLORS.BLUE_MYC};
-  cursor: pointer;
-  font-weight: bold;
-  &:hover {
-    color: ${COLORS.BLUE_LIGHT_DARKISH};
-  }
-`;
 
 // const WalletService = WalletFactory(WalletId.LEDGER_NANO_S);
 
@@ -114,7 +50,6 @@ const LedgerDecrypt = ({ formData, onUnlock }: OwnProps) => {
   }));
   const baseAsset = getAssetByUUID(assets)(network.baseAsset) as ExtendedAsset;
   const [assetToUse, setAssetToUse] = useState(baseAsset);
-  const [dpathAddView, setDpathAddView] = useState(false);
   const {
     state,
     requestConnection,
@@ -122,55 +57,18 @@ const LedgerDecrypt = ({ formData, onUnlock }: OwnProps) => {
     addDPaths,
     generateFreshAddress
   } = useDeterministicWallet(extendedDPaths, WalletId.LEDGER_NANO_S_NEW, DEFAULT_GAP_TO_SCAN_FOR);
-  const [freshAddressIndex, setFreshAddressIndex] = useState(0);
+
   // @todo -> Figure out which assets to display in dropdown. Dropdown is heavy with 900+ assets in it. Loads slow af.
   const filteredAssets = assets.filter(({ uuid }) => MOONPAY_ASSET_UUIDS.includes(uuid)); // @todo - fix this.
-
-  const handleNullConnect = () => {
-    requestConnection(network, assetToUse);
-  };
 
   const handleAssetUpdate = (newAsset: ExtendedAsset) => {
     setAssetToUse(newAsset);
     updateAsset(newAsset);
   };
 
-  const handleDPathAddition = (values: FormValues) => {
-    addDPaths([
-      {
-        ...values,
-        offset: 0,
-        numOfAddresses: numOfAccountsToCheck
-      }
-    ]);
-    setDpathAddView(false);
+  const handleNullConnect = () => {
+    requestConnection(network, assetToUse);
   };
-
-  interface FormValues {
-    label: string;
-    value: string;
-  }
-
-  const initialFormikValues: FormValues = {
-    label: '',
-    value: ''
-  };
-
-  const handleFreshAddressGeneration = () => {
-    if (freshAddressIndex > DEFAULT_GAP_TO_SCAN_FOR || !state.completed) {
-      return;
-    }
-    const freshAddressGenerationSuccess = generateFreshAddress({
-      ...defaultDPath,
-      offset: freshAddressIndex,
-      numOfAddresses: 1
-    });
-    if (freshAddressGenerationSuccess) {
-      setFreshAddressIndex(freshAddressIndex + 1);
-    }
-  };
-
-  const handleDownload = () => window.open(accountsToCSV(state.finishedAccounts, assetToUse));
 
   if (!network) {
     // @todo: make this better.
@@ -193,82 +91,19 @@ const LedgerDecrypt = ({ formData, onUnlock }: OwnProps) => {
   }
 
   if (state.isConnected && state.asset && (state.queuedAccounts || state.finishedAccounts)) {
-    return dpathAddView ? (
-      <MnemonicWrapper>
-        <HeadingWrapper>
-          <Icon type="back" width={20} onClick={() => setDpathAddView(false)} />
-          <Title fontSize="32px" bold={true}>
-            <Trans id="DETERMINISTIC_CUSTOM_TITLE" />
-          </Title>
-        </HeadingWrapper>
-        <Typography>{translate('DETERMINISTIC_CUSTOM_GET_INFORMED')}</Typography>
-        <Formik initialValues={initialFormikValues} onSubmit={handleDPathAddition}>
-          {({ handleChange, values }) => (
-            <SForm>
-              <SLabel htmlFor="label">
-                <Trans id="DETERMINISTIC_CUSTOM_LABEL" />
-              </SLabel>
-              <Input
-                placeholder="Custom Path"
-                name="label"
-                value={values.label}
-                onChange={handleChange}
-              />
-              <SLabel htmlFor="value">
-                <Trans id="DETERMINISTIC_CUSTOM_LABEL_DPATH" />
-              </SLabel>
-              <Input
-                placeholder="m/44’/60’/0’/0’"
-                name="value"
-                value={values.value}
-                onChange={handleChange}
-              />
-              <SButton onClick={() => handleDPathAddition(values)}>
-                <Trans id="DETERMINISTIC_ADD_DPATH" />
-              </SButton>
-            </SForm>
-          )}
-        </Formik>
-        <Typography>
-          <Trans id="DETERMINISTIC_SEE_SUMMARY" />{' '}
-          <SLink onClick={handleDownload}>
-            <Trans id="DETERMINISTIC_ALTERNATIVES_5" />
-          </SLink>
-          .
-        </Typography>
-      </MnemonicWrapper>
-    ) : (
-      <MnemonicWrapper>
-        <Title fontSize="32px" bold={true}>
-          <Trans id="MNEMONIC_TITLE" />
-        </Title>
-        <Typography>
-          <Trans id="MNEMONIC_SUBTITLE" />
-        </Typography>
-        <Parameters>
-          <AssetSelector
-            selectedAsset={assetToUse}
-            assets={filteredAssets}
-            onSelect={(option: ExtendedAsset) => {
-              handleAssetUpdate(option);
-            }}
-          />
-          <Button onClick={() => setDpathAddView(true)} inverted={true}>
-            <Trans id="MNEMONIC_ADD_CUSTOM_DPATH" />
-          </Button>
-        </Parameters>
-        <TableContainer>
-          <DeterministicAccountList
-            isComplete={state.completed}
-            asset={state.asset}
-            finishedAccounts={state.finishedAccounts}
-            onUnlock={onUnlock}
-            network={network}
-            generateFreshAddress={handleFreshAddressGeneration}
-            handleUpdate={updateAsset}
-          />
-        </TableContainer>
-      </MnemonicWrapper>
+    return (
+      <DeterministicWallet
+        state={state}
+        defaultDPath={defaultDPath}
+        assets={filteredAssets}
+        assetToUse={assetToUse}
+        network={network}
+        updateAsset={updateAsset}
+        addDPaths={addDPaths}
+        generateFreshAddress={generateFreshAddress}
+        handleAssetUpdate={handleAssetUpdate}
+        onUnlock={onUnlock}
+      />
     );
   } else {
     return (
